remove Select helper
